+++ frontend/src/features/workspace/components/AdminFeedbackContainer/index.ts
@@ -1,0 +1,1 @@
+export { AdminFeedbackContainer as default } from './AdminFeedbackContainer'
